Remove obsolete source files and add response code constants
Deleted several unused source files including enums, abstracts, and modules in the 'fairylandfuture' directory. Introduced a new 'response' package with response code mappings for handling HTTP response codes.
