+++ docs/zh-CN/manuals/modules/datetimes.md
@@ -23,7 +23,7 @@ from fairylandfuture.modules.datetimes import DateTimeModule
 ### 获取当前日期
 
 ```python
-from fairylandfuture.const.enums import DateTimeEnum
+from fairylandfuture.enums.datetimes import DateTimeEnum
 from fairylandfuture.modules.datetimes import DateTimeModule
 
 date = DateTimeModule.date()  # default format:: %Y-%m-%d
