+++ fairylandfuture/enums/files.py
@@ -1,0 +1,33 @@
+# coding: utf-8
+"""
+@software: PyCharm
+@author: Lionel Johnson
+@contact: https://fairy.host
+@organization: https://github.com/FairylandFuture
+@since: 2024-08-18 00:57:31 UTC+08:00
+"""
+
+from fairylandfuture.core.superclass.enumerate import BaseEnum
+
+
+class FileModeEnum(BaseEnum):
+    """
+    file mode enum.
+    """
+
+    r = "r"
+    rb = "rb"
+    r_plus = "r+"
+    rb_plus = "rb+"
+
+    w = "w"
+    wb = "wb"
+    w_plus = "w+"
+    wb_plus = "wb+"
+
+    a = "a"
+    ab = "ab"
+    a_plus = "a+"
+    ab_plus = "ab+"
+
+    default = r
