+++ fairylandfuture/enums/journal.py
@@ -1,0 +1,26 @@
+# coding: utf-8
+"""
+@software: PyCharm
+@author: Lionel Johnson
+@contact: https://fairy.host
+@organization: https://github.com/FairylandFuture
+@since: 2024-08-18 00:56:51 UTC+08:00
+"""
+
+from fairylandfuture.core.superclass.enumerate import BaseEnum
+
+
+class LogLevelEnum(BaseEnum):
+    """
+    Log level Enum.
+    """
+
+    trace = "TRACE"
+    debug = "DEBUG"
+    info = "INFO"
+    success = "SUCCESS"
+    warning = "WARNING"
+    error = "ERROR"
+    critical = "CRITICAL"
+
+    default = info
