+++ fairylandfuture/interface/__init__.py
@@ -4,5 +4,5 @@
 @author: Lionel Johnson
 @contact: https://fairy.host
 @organization: https://github.com/FairylandFuture
-@since: 2024-05-10 10:45:02 UTC+08:00
+@since: 2024-08-18 01:03:44 UTC+08:00
 """
