+++ fairylandfuture/structures/general/api.py
@@ -8,10 +8,10 @@
 """
 
 from dataclasses import dataclass, field
+from typing import MutableSequence, Sequence, MutableMapping, Mapping, Union
 
 from fairylandfuture.core.superclass.structures import BaseStructure
-
-from typing import MutableSequence, Sequence, MutableMapping, Mapping, Union
+from fairylandfuture.const.response.code import RESPONSE_CODE_MAP
 
 
 @dataclass(frozen=True)
@@ -20,72 +20,9 @@ class ResponseStructure(BaseStructure):
     message: str = field(default=None)
     data: Union[MutableSequence, Sequence, MutableMapping, Mapping] = field(default=None)
 
-    def __default_message(self, _k):
-        response_code_mapping = {
-            100: "Continue",
-            101: "Switching Protocols",
-            102: "Processing",
-            200: "OK",
-            201: "Created",
-            202: "Accepted",
-            203: "Non-Authoritative Information",
-            204: "No Content",
-            205: "Reset Content",
-            206: "Partial Content",
-            207: "Multi-Status",
-            300: "Multiple Choices",
-            301: "Moved Permanently",
-            302: "Found",
-            303: "See Other",
-            304: "Not Modified",
-            307: "Temporary Redirect",
-            308: "Permanent Redirect",
-            400: "Bad Request",
-            401: "Unauthorized",
-            402: "Payment Required",
-            403: "Forbidden",
-            404: "Not Found",
-            405: "Method Not Allowed",
-            406: "Not Acceptable",
-            407: "Proxy Authentication Required",
-            408: "Request Timeout",
-            409: "Conflict",
-            410: "Gone",
-            411: "Length Required",
-            412: "Precondition Failed",
-            413: "Payload Too Large",
-            414: "URI Too Long",
-            415: "Unsupported Media Type",
-            416: "Range Not Satisfiable",
-            417: "Expectation Failed",
-            418: "I'm a teapot",
-            421: "Misdirected Request",
-            422: "Unprocessable Entity",
-            423: "Locked",
-            424: "Failed Dependency",
-            425: "Too Early",
-            426: "Upgrade Required",
-            428: "Precondition Required",
-            429: "Too Many Requests",
-            431: "Request Header Fields Too Large",
-            451: "Unavailable For Legal Reasons",
-            500: "Internal Server Error",
-            501: "Not Implemented",
-            502: "Bad Gateway",
-            503: "Service Unavailable",
-            504: "Gateway Timeout",
-            505: "HTTP Version Not Supported",
-            506: "Variant Also Negotiates",
-            507: "Insufficient Storage",
-            508: "Loop Detected",
-            510: "Not Extended",
-            511: "Network Authentication Required",
-        }
-        return response_code_mapping.get(_k, "Internal Server Error")
-
     def __post_init__(self):
         if self.code and not self.message:
-            object.__setattr__(self, "message", self.__default_message(self.code))
+            object.__setattr__(self, "message", RESPONSE_CODE_MAP.get(self.code, "Internal Server Error"))
 
     def __str__(self):
         return self.string
