+++ test/devunit/archive/20240512/_test_20240512.py
@@ -13,7 +13,7 @@ from datetime import datetime
 from _test import TestBase
 from dateutil.relativedelta import relativedelta
 
-from fairylandfuture.const.enums import DateTimeEnum
+from fairylandfuture.enums.datetimes import DateTimeEnum
 from fairylandfuture.modules.datetimes import DateTimeModule
 from fairylandfuture.modules.decorators import ActionDecorator, SingletonDecorator, TimingDecorator, TipsDecorator
 from fairylandfuture.utils.journal import journal
