+++ test/devunit/archive/20240519/_test.py
@@ -13,7 +13,7 @@ from typing import Any, Dict
 import yaml
 
 from bin.general import DEV_CASE_CONFIG_FILE_PATH
-from fairylandfuture.const.enums import DateTimeEnum, EncodingEnum
+from fairylandfuture.enums.datetimes import DateTimeEnum, EncodingEnum
 from fairylandfuture.structures.builder.databases import StructureMySQLExecute, StructureSQLInsertManyParams
 from fairylandfuture.modules.datasource import MySQLDataSource
