+++ test/devunit/archive/20240601/_test.py
@@ -7,7 +7,7 @@
 @since: 2024-06-02 12:02:50 UTC+08:00
 """
 
-from fairylandfuture.const.enums import DateTimeEnum
+from fairylandfuture.enums.datetimes import DateTimeEnum
 from fairylandfuture.modules.datetimes import DateTimeModule
 from fairylandfuture.utils.journal import journal
 from fairylandfuture.utils.networking.local import LocalNetworkUtils
