+++ test/devunit/archive/20240603/_test.py
@@ -8,7 +8,7 @@
 """
 
 from fairylandfuture.modules.datetimes import DateTimeModule
-from fairylandfuture.const.enums import DateTimeEnum
+from fairylandfuture.enums.datetimes import DateTimeEnum
 
 a = DateTimeModule.datetime_to_timestamp("2024-01-01", _format=DateTimeEnum.date)
 a = DateTimeModule.datetime_to_timestamp("2024/01/01", _format="%Y/%m/%d")
