+++ test/devunit/archive/20240614/_test.py
@@ -9,7 +9,7 @@
 
 import binascii
 
-from fairylandfuture.const.enums import EncodingEnum
+from fairylandfuture.enums.enconding import EncodingEnum
 from fairylandfuture.modules.encryptions import UserPasswordEncryption
 from fairylandfuture.modules.encryptions import Base64Encryption
